+++ RELEASE_NOTES.md
@@ -102,6 +102,12 @@ Examples
 * UPDATED: Most examples that use parametric entities to use parameter
 variables.
 
+Installers and installation scripts
+-----------------------------------
+
+* FIXED: The manual `install.sh` script to use `mkdir -p` when creating the
+Logtalk installation directory.
+
 
 3.44.0 - February 3, 2021
 =========================